Add some tests for OsQueue
